requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
chromadb
sentence-transformers
pypdf
groq
python-dotenv